# docs/architecture/podpay-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PodPay API",
    "version": "1.0.0",
...
    "contact": {
      "name": "PodPay Suporte",
      "email": "contato@podpay.app",
      "url": "https://podpay.app"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://www.podpay.app/docs/termos-e-condicoes"
    },
    "x-logo": {
      "url": "https://www.podpay.app/images/logos/PodPayHeader.svg",
      "altText": "PodPay Gateway",
      "backgroundColor": "#0d3323"
    }
  },
  "servers": [
    {
      "url": "https://api.podpay.app",
      "description": "Produção"
    },
    {
      "url": "https://sandbox.podpay.app",
      "description": "Sandbox (Testes)"
    },
    {
      "url": "/",
      "description": "Local (Desenvolvimento)"
    }
  ],
...
  "paths": {
...
    "/v1/transactions/{id}/refund": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Estornar venda",
        "description": "Estorna uma transação que foi paga. Apenas transações com status 'paid' podem ser estornadas. O valor será debitado do saldo disponível.",
        "operationId": "refundTransaction",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chave de API (obrigatoria). Formato sk_live_... ou sk_test_... (dashboard PodPay)."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da transação"
          }
        ],
        "responses": {
          "200": {
            "description": "Transação estornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TransactionApiResponse"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "tx_clx1234567890abcdef",
                    "status": "refunded",
                    "amount": 10000,
                    "paymentMethod": "pix",
                    "customer": {
                      "document": {
                        "type": "cpf",
                        "number": "12345678900"
                      },
                      "name": "João Silva",
                      "email": "joao@example.com",
                      "phone": "11999999999"
                    },
                    "createdAt": "2024-01-15T10:30:00.000Z"
                  },
                  "meta": {
                    "timestamp": "2024-01-15T10:40:00.000Z",
                    "version": "v1",
                    "requestId": "req_1705316400000_refund123"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notPaid": {
                    "summary": "Transação não está paga",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "TRANSACTION_NOT_PAID",
                        "message": "Apenas transações pagas podem ser estornadas"
                      },
                      "meta": {
                        "timestamp": "2024-01-15T10:40:00.000Z",
                        "version": "v1",
                        "requestId": "req_1705316400000_refund123"
                      }
                    }
                  },
                  "alreadyRefunded": {
                    "summary": "Transação já foi estornada",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "TRANSACTION_ALREADY_REFUNDED",
                        "message": "Transação já foi estornada"
                      },
                      "meta": {
                        "timestamp": "2024-01-15T10:40:00.000Z",
                        "version": "v1",
                        "requestId": "req_1705316400000_refund123"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v1/withdrawals/{id}": {
      "get": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Buscar saque por ID",
        "description": "Busca um saque específico pelo seu ID.",
        "operationId": "getWithdrawal",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chave de API (obrigatoria). Formato sk_live_... ou sk_test_... (dashboard PodPay)."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do saque"
          }
        ],
        "responses": {
          "200": {
            "description": "Saque encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WithdrawalApiResponse"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "wd_clx9876543210fedcba",
                    "method": "fiat",
                    "status": "completed",
                    "amount": 50000,
                    "fee": 500,
                    "netAmount": 49500,
                    "createdAt": "2024-01-15T11:00:00.000Z",
                    "completedAt": "2024-01-15T11:05:00.000Z"
                  },
                  "meta": {
                    "timestamp": "2024-01-15T11:10:00.000Z",
                    "version": "1.0"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v1/withdrawals/{id}/cancel": {
      "patch": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Cancelar saque",
        "description": "Cancela um saque que está pendente. Apenas saques com status 'pending' podem ser cancelados.",
        "operationId": "cancelWithdrawal",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chave de API (obrigatoria). Formato sk_live_... ou sk_test_... (dashboard PodPay)."
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do saque"
          }
        ],
        "responses": {
          "200": {
            "description": "Saque cancelado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/WithdrawalApiResponse"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "success": true,
                  "data": {
                    "id": "wd_clx9876543210fedcba",
                    "method": "fiat",
                    "status": "canceled",
                    "amount": 50000,
                    "fee": 500,
                    "netAmount": 49500,
                    "createdAt": "2024-01-15T11:00:00.000Z"
                  },
                  "meta": {
                    "timestamp": "2024-01-15T11:15:00.000Z",
                    "version": "1.0"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key para autenticacao REST. Use `sk_*` para a API principal ou `ck_*` apenas para Checkout hospedado. Obtenha as chaves no dashboard do PodPay."
      }
    },
    "schemas": {
...
      "CustomerInfo": {
        "type": "object",
        "required": [
          "document",
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "document": {
            "type": "object",
            "required": [
              "type",
              "number"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "cpf",
                  "cnpj"
                ],
                "description": "Tipo de documento"
              },
              "number": {
                "type": "string",
                "description": "Número do documento (apenas dígitos)"
              }
            }
          },
          "name": {
            "type": "string",
            "description": "Nome completo do cliente"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do cliente"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do cliente (formato: DDD + número, ex: 11999999999)"
          }
        }
      },
...
      "TransactionApiResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da transação"
          },
          "companyId": {
            "type": "string",
            "description": "ID da empresa"
          },
          "currency": {
            "type": "string",
            "enum": [
              "BRL"
            ],
            "description": "Moeda (sempre BRL)"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "PAID",
              "FAILED",
              "CANCELED",
              "BLOCKED",
              "REFUNDED",
              "PRE_CHARGEBACK",
              "CHARGEBACK"
            ],
            "description": "Status da transação (UPPERCASE)"
          },
          "amount": {
            "type": "integer",
            "description": "Valor em centavos"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "pix",
              "credit_card",
              "boleto"
            ],
            "description": "Método de pagamento"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerInfo",
            "description": "Dados do cliente"
          },
          "pixQrCode": {
            "type": "string",
            "description": "Código PIX (copy-paste) - apenas para PIX"
          },
          "pixQrCodeImage": {
            "type": "string",
            "format": "uri",
            "description": "URL da imagem do QR Code PIX - apenas para PIX"
          },
          "boletoUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL do boleto - apenas para Boleto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação (ISO 8601)"
          }
        }
      },
...
      "WithdrawalApiResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do saque"
          },
          "method": {
            "type": "string",
            "enum": [
              "fiat",
              "crypto"
            ],
            "description": "Tipo de saque"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "pending_approval",
              "processing",
              "completed",
              "failed",
              "canceled"
            ],
            "description": "Status do saque"
          },
          "amount": {
            "type": "integer",
            "description": "Valor solicitado em centavos"
          },
          "fee": {
            "type": "integer",
            "description": "Taxa em centavos"
          },
          "netAmount": {
            "type": "integer",
            "description": "Valor líquido em centavos"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação (ISO 8601)"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de conclusão (apenas para status=completed)"
          },
          "failureReason": {
            "type": "string",
            "description": "Motivo da falha (apenas para status=failed)"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "description": "Formato padrão de resposta da API",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica se a requisição foi bem-sucedida"
          },
          "data": {
            "description": "Dados da resposta (estrutura varia conforme o endpoint)"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorResponse",
            "description": "Objeto de erro (apenas quando success=false)"
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "description": "Timestamp da resposta (ISO 8601)"
              },
              "version": {
                "type": "string",
                "description": "Versão da API (ex: v1, 1.0)"
              },
              "requestId": {
                "type": "string",
                "description": "ID único da requisição para rastreamento (formato: req_${timestamp}_${random})"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard API error object. Fields `traceId`, `timestamp`, `path`, and `details` are optional and additive for integrators. Messages are provider-agnostic (no acquirer/PSP brand names); use `error.code`, `details`, and `traceId` when contacting support.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. VALIDATION_ERROR, AUTHENTICATION_ERROR, DUPLICATE_WITHDRAWAL_RISK)"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message without acquirer or PSP identifiers"
          },
          "details": {
            "description": "Optional structured payload: validation field list, `retryAfter` seconds, `existingWithdrawalId` for duplicate withdrawal risk, etc. Shape varies by `code`."
          },
          "integrationError": {
            "description": "Optional structured integration hint (transactions). Same additive rules as `details`."
          },
          "traceId": {
            "type": "string",
            "description": "Correlation / trace identifier for support and logs"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "When the error was materialized (ISO 8601)"
          },
          "path": {
            "type": "string",
            "description": "Request path that produced the error"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Erro de validação",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  }
                }
              ]
            },
            "example": {
              "success": false,
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Dados inválidos",
                "details": [
                  {
                    "field": "amount",
                    "message": "Valor mínimo é R$ 1,00 (100 centavos)"
                  }
                ]
              },
              "meta": {
                "timestamp": "2024-01-15T12:00:00.000Z",
                "version": "v1",
                "requestId": "req_1705318800000_error123"
              }
            }
          }
        }
      },
      "AuthenticationError": {
        "description": "Erro de autenticação",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  }
                }
              ]
            },
            "example": {
              "success": false,
              "error": {
                "code": "AUTHENTICATION_ERROR",
                "message": "API key inválida ou ausente"
              },
              "meta": {
                "timestamp": "2024-01-15T12:00:00.000Z",
                "version": "v1",
                "requestId": "req_1705318800000_auth123"
              }
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate limit excedido",
        "headers": {
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            },
            "description": "Limite de requisições"
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Requisições restantes"
          },
          "X-RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Timestamp de reset (Unix timestamp)"
          },
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Segundos para tentar novamente"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  }
                }
              ]
            },
            "example": {
              "success": false,
              "error": {
                "code": "RATE_LIMIT_EXCEEDED",
                "message": "Rate limit exceeded. Please try again later."
              },
              "meta": {
                "timestamp": "2024-01-15T12:00:00.000Z",
                "version": "v1",
                "requestId": "req_1705318800000_rate123"
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  }
                }
              ]
            },
            "example": {
              "success": false,
              "error": {
                "code": "TRANSACTION_NOT_FOUND",
                "message": "Transação não encontrada"
              },
              "meta": {
                "timestamp": "2024-01-15T12:00:00.000Z",
                "version": "v1",
                "requestId": "req_1705318800000_notfound123"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  }
                }
              ]
            },
            "example": {
              "success": false,
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "message": "Erro interno do servidor"
              },
              "meta": {
                "timestamp": "2024-01-15T12:00:00.000Z",
                "version": "v1",
                "requestId": "req_1705318800000_server123"
              }
            }
          }
        }
      }
    }
  },
...
}